move InvokeAsync to the new HelperProcess
