+++ src/shared/Microsoft.Git.CredentialManager/HelperProcess.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license.
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Threading.Tasks;
 
 namespace Microsoft.Git.CredentialManager
@@ -31,7 +32,42 @@ namespace Microsoft.Git.CredentialManager
 
         public async Task<IDictionary<string, string>> InvokeAsync(string path, string args, IDictionary<string, string> standardInput = null)
         {
-            throw new NotImplementedException();
+            var procStartInfo = new ProcessStartInfo(path)
+            {
+                Arguments = args,
+                RedirectStandardInput = true,
+                RedirectStandardOutput = true,
+                RedirectStandardError = false, // Do not redirect stderr as tracing might be enabled
+                UseShellExecute = false
+            };
+
+            var process = Process.Start(procStartInfo);
+            if (process is null)
+            {
+                throw new Exception($"Failed to start helper process '{path}'");
+            }
+
+            if (!(standardInput is null))
+            {
+                await process.StandardInput.WriteDictionaryAsync(standardInput);
+            }
+
+            IDictionary<string, string> resultDict = await process.StandardOutput.ReadDictionaryAsync(StringComparer.OrdinalIgnoreCase);
+
+            await Task.Run(() => process.WaitForExit());
+            int exitCode = process.ExitCode;
+
+            if (exitCode != 0)
+            {
+                if (!resultDict.TryGetValue("error", out string errorMessage))
+                {
+                    errorMessage = "Unknown";
+                }
+
+                throw new Exception($"helper error ({exitCode}): {errorMessage}");
+            }
+
+            return resultDict;
         }
     }
 }
